minor: refactor transactions example
